pyproject.toml
[project]
name = "anygraph-backend"
version = "1.0.0"
description = "Backend for AnyGraph - Easy data analysis platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.123.0",
    "uvicorn>=0.30.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "docker>=7.0.0",
    "google-genai>=1.0.0",
    "openpyxl>=3.1.0",
    "pydantic[email]>=2.0.0",
    "tabulate>=0.9.0",
]
